Создал наследника для WebView
